Refactory na api para padrão repository

+++ api/src/routes/creditCard.js
@@ -1,22 +1,23 @@
-import { CreditCard } from '../sequelize/models'
 import { createCreditCardService } from '../services'
 import errorHandler from '../util/errorHandler'
+import { CreditCardRepository } from '../repository'
 
-const repository = {
-  getAll: (userId) => CreditCard.findAll({ where: { userId: userId } }),
-  create: (card) => CreditCard.create(card),
-  update: (card, userId) => CreditCard.update({ name: card.name }, { where: { id: card.id, userId: userId } }),
-  remove: (id, userId) => CreditCard.destroy({ where: { id: id, userId: userId } })
-}
-
-const service = createCreditCardService(repository)
+const service = createCreditCardService(CreditCardRepository)
 
 export default (app) => {
-  app.get('/api/credit-card', errorHandler((req, res) => service.getAll(req.claims.id).then(result => res.json(result))))
+  app.get('/api/credit-card', errorHandler((req, res) => service.getByUser(req.claims.id).then(result => res.json(result))))
 
-  app.post('/api/credit-card', errorHandler((req, res) => service.create(req.body, req.claims.id).then(result => res.json(result))))
+  app.post('/api/credit-card', errorHandler(async (req, res) => {
+    const card = req.body || {}
+    card.userId = req.claims.id
+    return service.create(card).then(result => res.json(result))
+  }))
 
-  app.put('/api/credit-card', errorHandler((req, res) => service.update(req.body, req.claims.id).then(result => res.json(result))))
+  app.put('/api/credit-card', errorHandler(async (req, res) => {
+    const card = req.body || {}
+    card.userId = req.claims.id
+    return service.update(card).then(result => res.json(result))
+  }))
 
-  app.delete('/api/credit-card/:id', errorHandler((req, res) => service.remove(req.params.id, req.claims.id).then(result => res.json(result))))
+  app.delete('/api/credit-card/:id', errorHandler((req, res) => service.remove({ id: req.params.id, userId: req.claims.id }).then(result => res.json(result))))
 }
